requirements.txt
# RAG
weaviate-client
rank-bm25
FlagEmbedding
docling
llama-index

# HTTP Server
fastapi[standard]
uvicorn

# Shared
pydantic
dotenv

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0
httpx>=0.24.0
faker>=18.0.0